==> TMB_TLR9/0_MAF_Concat.ipynb <==
# Generate MAF files by Concat

import os
import pandas as pd

### 1. Get folder list

folder_lists = os.listdir('./MAF/')

### 2. Generate backbone dataframe using 1st set and remove its name in the folder list

col = [x for x in range(0, 139)]
Maf_files=pd.read_csv('./MAF/'+folder_lists[0]+'/'+os.listdir('./MAF/'+folder_lists[0])[0],
                    header=7, usecols=col, sep='\t')
del folder_lists[0]

## 3. Iterate folders using for loops

# '.wxs.aliquot_ensemble_masked.maf.gz'로 끝나는 파일을 찾아 이어 붙인다. 
for folder in folder_lists:
    f_names=os.listdir('./MAF/'+folder+'/')
    for f_name in f_names:
        if f_name.endswith('.wxs.aliquot_ensemble_masked.maf.gz'):
            df_a= pd.read_csv('./MAF/'+folder+'/'+f_name, header=7, usecols=col, sep='\t')
            Maf_files=pd.concat([Maf_files, df_a])

### 4. Save files

Maf_files.to_csv('./Maf_files.tsv', sep='\t', header=True, index=False)
